+++ src/main/java/fr/cryptohash/DigestEngine.java
@@ -150,7 +150,7 @@ public abstract class DigestEngine implements Digest {
 	/** @see Digest */
 	public void update(byte input)
 	{
-		inputBuf[inputLen ++] = (byte)input;
+		inputBuf[inputLen ++] = input;
 		if (inputLen == blockLen) {
 			processBlock(inputBuf);
 			blockCount ++;
