+++ src/main/java/fr/cryptohash/MD5.java
@@ -74,10 +74,10 @@ public class MD5 extends MDHelper {
 	/** @see DigestEngine */
 	protected void engineReset()
 	{
-		currentVal[0] = (int)0x67452301;
-		currentVal[1] = (int)0xEFCDAB89;
-		currentVal[2] = (int)0x98BADCFE;
-		currentVal[3] = (int)0x10325476;
+		currentVal[0] = 0x67452301;
+		currentVal[1] = 0xEFCDAB89;
+		currentVal[2] = 0x98BADCFE;
+		currentVal[3] = 0x10325476;
 	}
 
 	/** @see DigestEngine */
